Move eth config and builder ctx to reth-rpc-eth-types & make RpcRegistry generic over EthApi
