+++ crates/rpc/rpc-builder/src/eth.rs
@@ -1,30 +1,10 @@
-use std::{fmt::Debug, time::Duration};
-
 use reth_evm::ConfigureEvm;
-use reth_provider::{
-    BlockReader, BlockReaderIdExt, CanonStateSubscriptions, ChainSpecProvider, EvmEnvProvider,
-    StateProviderFactory,
-};
-use reth_rpc::{eth::EthFilterConfig, EthFilter, EthPubSub};
+use reth_provider::{BlockReader, CanonStateSubscriptions, EvmEnvProvider, StateProviderFactory};
+use reth_rpc::{EthFilter, EthPubSub};
 use reth_rpc_eth_types::{
-    cache::cache_new_blocks_task, fee_history::fee_history_cache_new_blocks_task, EthStateCache,
-    EthStateCacheConfig, FeeHistoryCache, FeeHistoryCacheConfig, GasPriceOracle,
-    GasPriceOracleConfig, RPC_DEFAULT_GAS_CAP,
-};
-use reth_rpc_server_types::constants::{
-    default_max_tracing_requests, DEFAULT_ETH_PROOF_WINDOW, DEFAULT_MAX_BLOCKS_PER_FILTER,
-    DEFAULT_MAX_LOGS_PER_RESPONSE, DEFAULT_PROOF_PERMITS,
+    cache::cache_new_blocks_task, EthApiBuilderCtx, EthConfig, EthStateCache,
 };
 use reth_tasks::TaskSpawner;
-use serde::{Deserialize, Serialize};
-
-/// Default value for stale filter ttl
-const DEFAULT_STALE_FILTER_TTL: Duration = Duration::from_secs(5 * 60);
-
-/// Alias for function that builds the core `eth` namespace API.
-pub type DynEthApiBuilder<Provider, Pool, EvmConfig, Network, Tasks, Events, EthApi> =
-    Box<dyn FnOnce(&EthApiBuilderCtx<Provider, Pool, EvmConfig, Network, Tasks, Events>) -> EthApi>;
-
 /// Handlers for core, filter and pubsub `eth` namespace APIs.
 #[derive(Debug, Clone)]
 pub struct EthHandlers<Provider, Pool, Network, Events, EthApi> {
@@ -41,7 +21,7 @@ pub struct EthHandlers<Provider, Pool, Network, Events, EthApi> {
 impl<Provider, Pool, Network, Events, EthApi> EthHandlers<Provider, Pool, Network, Events, EthApi> {
     /// Returns a new [`EthHandlers`] builder.
     #[allow(clippy::too_many_arguments)]
-    pub fn builder<EvmConfig, Tasks, EthApiB>(
+    pub fn builder<EvmConfig, Tasks>(
         provider: Provider,
         pool: Pool,
         network: Network,
@@ -49,12 +29,10 @@ impl<Provider, Pool, Network, Events, EthApi> EthHandlers<Provider, Pool, Networ
         config: EthConfig,
         executor: Tasks,
         events: Events,
-        eth_api_builder: EthApiB,
-    ) -> EthHandlersBuilder<Provider, Pool, Network, Tasks, Events, EvmConfig, EthApi>
-    where
-        EthApiB: FnOnce(&EthApiBuilderCtx<Provider, Pool, EvmConfig, Network, Tasks, Events>) -> EthApi
-            + 'static,
-    {
+        eth_api_builder: Box<
+            dyn EthApiBuilder<Provider, Pool, EvmConfig, Network, Tasks, Events, Output = EthApi>,
+        >,
+    ) -> EthHandlersBuilder<Provider, Pool, Network, Tasks, Events, EvmConfig, EthApi> {
         EthHandlersBuilder {
             provider,
             pool,
@@ -63,7 +41,7 @@ impl<Provider, Pool, Network, Events, EthApi> EthHandlers<Provider, Pool, Networ
             config,
             executor,
             events,
-            eth_api_builder: Box::new(eth_api_builder),
+            eth_api_builder,
         }
     }
 }
@@ -78,7 +56,8 @@ pub struct EthHandlersBuilder<Provider, Pool, Network, Tasks, Events, EvmConfig,
     config: EthConfig,
     executor: Tasks,
     events: Events,
-    eth_api_builder: DynEthApiBuilder<Provider, Pool, EvmConfig, Network, Tasks, Events, EthApi>,
+    eth_api_builder:
+        Box<dyn EthApiBuilder<Provider, Pool, EvmConfig, Network, Tasks, Events, Output = EthApi>>,
 }
 
 impl<Provider, Pool, Network, Tasks, Events, EvmConfig, EthApi>
@@ -87,9 +66,10 @@ where
     Provider: StateProviderFactory + BlockReader + EvmEnvProvider + Clone + Unpin + 'static,
     Pool: Send + Sync + Clone + 'static,
     EvmConfig: ConfigureEvm,
-    Network: Clone,
+    Network: Clone + 'static,
     Tasks: TaskSpawner + Clone + 'static,
-    Events: CanonStateSubscriptions + Clone,
+    Events: CanonStateSubscriptions + Clone + 'static,
+    EthApi: 'static,
 {
     /// Returns a new instance with handlers for `eth` namespace.
     pub fn build(self) -> EthHandlers<Provider, Pool, Network, Events, EthApi> {
@@ -123,7 +103,7 @@ where
             cache,
         };
 
-        let api = eth_api_builder(&ctx);
+        let api = eth_api_builder.build_eth_api(&ctx);
 
         let filter = EthFilterApiBuilder::build(&ctx);
 
@@ -133,150 +113,37 @@ where
     }
 }
 
-/// Additional config values for the eth namespace.
-#[derive(Debug, Clone, Eq, PartialEq, Serialize, Deserialize)]
-pub struct EthConfig {
-    /// Settings for the caching layer
-    pub cache: EthStateCacheConfig,
-    /// Settings for the gas price oracle
-    pub gas_oracle: GasPriceOracleConfig,
-    /// The maximum number of blocks into the past for generating state proofs.
-    pub eth_proof_window: u64,
-    /// The maximum number of tracing calls that can be executed in concurrently.
-    pub max_tracing_requests: usize,
-    /// Maximum number of blocks that could be scanned per filter request in `eth_getLogs` calls.
-    pub max_blocks_per_filter: u64,
-    /// Maximum number of logs that can be returned in a single response in `eth_getLogs` calls.
-    pub max_logs_per_response: usize,
-    /// Gas limit for `eth_call` and call tracing RPC methods.
-    ///
-    /// Defaults to [`RPC_DEFAULT_GAS_CAP`]
-    pub rpc_gas_cap: u64,
-    ///
-    /// Sets TTL for stale filters
-    pub stale_filter_ttl: Duration,
-    /// Settings for the fee history cache
-    pub fee_history_cache: FeeHistoryCacheConfig,
-    /// The maximum number of getproof calls that can be executed concurrently.
-    pub proof_permits: usize,
-}
-
-impl EthConfig {
-    /// Returns the filter config for the `eth_filter` handler.
-    pub fn filter_config(&self) -> EthFilterConfig {
-        EthFilterConfig::default()
-            .max_blocks_per_filter(self.max_blocks_per_filter)
-            .max_logs_per_response(self.max_logs_per_response)
-            .stale_filter_ttl(self.stale_filter_ttl)
-    }
-}
-
-impl Default for EthConfig {
-    fn default() -> Self {
-        Self {
-            cache: EthStateCacheConfig::default(),
-            gas_oracle: GasPriceOracleConfig::default(),
-            eth_proof_window: DEFAULT_ETH_PROOF_WINDOW,
-            max_tracing_requests: default_max_tracing_requests(),
-            max_blocks_per_filter: DEFAULT_MAX_BLOCKS_PER_FILTER,
-            max_logs_per_response: DEFAULT_MAX_LOGS_PER_RESPONSE,
-            rpc_gas_cap: RPC_DEFAULT_GAS_CAP.into(),
-            stale_filter_ttl: DEFAULT_STALE_FILTER_TTL,
-            fee_history_cache: FeeHistoryCacheConfig::default(),
-            proof_permits: DEFAULT_PROOF_PERMITS,
-        }
-    }
-}
-
-impl EthConfig {
-    /// Configures the caching layer settings
-    pub const fn state_cache(mut self, cache: EthStateCacheConfig) -> Self {
-        self.cache = cache;
-        self
-    }
-
-    /// Configures the gas price oracle settings
-    pub const fn gpo_config(mut self, gas_oracle_config: GasPriceOracleConfig) -> Self {
-        self.gas_oracle = gas_oracle_config;
-        self
-    }
-
-    /// Configures the maximum number of tracing requests
-    pub const fn max_tracing_requests(mut self, max_requests: usize) -> Self {
-        self.max_tracing_requests = max_requests;
-        self
-    }
-
-    /// Configures the maximum block length to scan per `eth_getLogs` request
-    pub const fn max_blocks_per_filter(mut self, max_blocks: u64) -> Self {
-        self.max_blocks_per_filter = max_blocks;
-        self
-    }
-
-    /// Configures the maximum number of logs per response
-    pub const fn max_logs_per_response(mut self, max_logs: usize) -> Self {
-        self.max_logs_per_response = max_logs;
-        self
-    }
-
-    /// Configures the maximum gas limit for `eth_call` and call tracing RPC methods
-    pub const fn rpc_gas_cap(mut self, rpc_gas_cap: u64) -> Self {
-        self.rpc_gas_cap = rpc_gas_cap;
-        self
-    }
-
-    /// Configures the maximum proof window for historical proof generation.
-    pub const fn eth_proof_window(mut self, window: u64) -> Self {
-        self.eth_proof_window = window;
-        self
-    }
-
-    /// Configures the number of getproof requests
-    pub const fn proof_permits(mut self, permits: usize) -> Self {
-        self.proof_permits = permits;
-        self
-    }
-}
-
-/// Context for building the `eth` namespace API.
-#[derive(Debug, Clone)]
-pub struct EthApiBuilderCtx<Provider, Pool, EvmConfig, Network, Tasks, Events> {
-    /// Database handle.
-    pub provider: Provider,
-    /// Mempool handle.
-    pub pool: Pool,
-    /// Network handle.
-    pub network: Network,
-    /// EVM configuration.
-    pub evm_config: EvmConfig,
-    /// RPC config for `eth` namespace.
-    pub config: EthConfig,
-    /// Runtime handle.
-    pub executor: Tasks,
-    /// Events handle.
-    pub events: Events,
-    /// RPC cache handle.
-    pub cache: EthStateCache,
-}
-
-impl<Provider, Pool, EvmConfig, Network, Tasks, Events>
-    EthApiBuilderCtx<Provider, Pool, EvmConfig, Network, Tasks, Events>
-where
-    Provider: BlockReaderIdExt + Clone,
+/// Builds core `eth` namespace API to run on the RPC server
+/// ([`EthApiServer`](reth_rpc::eth::EthApiServer)).
+pub trait EthApiBuilder<Provider, Pool, Evm, Network, Tasks, Events>:
+    Send + Sync + Unpin + 'static
 {
-    /// Returns a new [`FeeHistoryCache`] for the context.
-    pub fn new_fee_history_cache(&self) -> FeeHistoryCache
-    where
-        Provider: ChainSpecProvider + 'static,
-        Tasks: TaskSpawner,
-        Events: CanonStateSubscriptions,
-    {
-        FeeHistoryCacheBuilder::build(self)
-    }
+    /// The `eth_` API type that is built.
+    type Output;
 
-    /// Returns a new [`GasPriceOracle`] for the context.
-    pub fn new_gas_price_oracle(&self) -> GasPriceOracle<Provider> {
-        GasPriceOracleBuilder::build(self)
+    /// Builds core `eth` namespace API.
+    fn build_eth_api(
+        &self,
+        ctx: &EthApiBuilderCtx<Provider, Pool, Evm, Network, Tasks, Events>,
+    ) -> Self::Output;
+}
+
+impl<Provider, Pool, Evm, Network, Tasks, Events, F, EthApi>
+    EthApiBuilder<Provider, Pool, Evm, Network, Tasks, Events> for F
+where
+    F: Fn(&EthApiBuilderCtx<Provider, Pool, Evm, Network, Tasks, Events>) -> EthApi
+        + Send
+        + Sync
+        + Unpin
+        + 'static,
+{
+    type Output = EthApi;
+
+    fn build_eth_api(
+        &self,
+        ctx: &EthApiBuilderCtx<Provider, Pool, Evm, Network, Tasks, Events>,
+    ) -> Self::Output {
+        self(ctx)
     }
 }
 
@@ -327,52 +194,5 @@ impl EthPubSubApiBuilder {
             ctx.network.clone(),
             Box::new(ctx.executor.clone()),
         )
-    }
-}
-
-/// Builds `eth_` core api component [`GasPriceOracle`], for given context.
-#[derive(Debug)]
-pub struct GasPriceOracleBuilder;
-
-impl GasPriceOracleBuilder {
-    /// Builds a [`GasPriceOracle`], for given context.
-    pub fn build<Provider, Pool, EvmConfig, Network, Tasks, Events>(
-        ctx: &EthApiBuilderCtx<Provider, Pool, EvmConfig, Network, Tasks, Events>,
-    ) -> GasPriceOracle<Provider>
-    where
-        Provider: BlockReaderIdExt + Clone,
-    {
-        GasPriceOracle::new(ctx.provider.clone(), ctx.config.gas_oracle, ctx.cache.clone())
-    }
-}
-
-/// Builds `eth_` core api component [`FeeHistoryCache`], for given context.
-#[derive(Debug)]
-pub struct FeeHistoryCacheBuilder;
-
-impl FeeHistoryCacheBuilder {
-    /// Builds a [`FeeHistoryCache`], for given context.
-    pub fn build<Provider, Pool, EvmConfig, Network, Tasks, Events>(
-        ctx: &EthApiBuilderCtx<Provider, Pool, EvmConfig, Network, Tasks, Events>,
-    ) -> FeeHistoryCache
-    where
-        Provider: ChainSpecProvider + BlockReaderIdExt + Clone + 'static,
-        Tasks: TaskSpawner,
-        Events: CanonStateSubscriptions,
-    {
-        let fee_history_cache =
-            FeeHistoryCache::new(ctx.cache.clone(), ctx.config.fee_history_cache);
-
-        let new_canonical_blocks = ctx.events.canonical_state_stream();
-        let fhc = fee_history_cache.clone();
-        let provider = ctx.provider.clone();
-        ctx.executor.spawn_critical(
-            "cache canonical blocks for fee history task",
-            Box::pin(async move {
-                fee_history_cache_new_blocks_task(fhc, new_canonical_blocks, provider).await;
-            }),
-        );
-
-        fee_history_cache
     }
 }
