+++ crates/rpc/rpc-builder/src/lib.rs
@@ -164,7 +164,7 @@ use reth_rpc_eth_api::{
     },
     EthApiServer, FullEthApiServer, RawTransactionForwarder,
 };
-use reth_rpc_eth_types::{EthStateCache, EthSubscriptionIdProvider};
+use reth_rpc_eth_types::{EthConfig, EthStateCache, EthSubscriptionIdProvider};
 use reth_rpc_layer::{AuthLayer, Claims, JwtAuthValidator, JwtSecret};
 use reth_tasks::{pool::BlockingTaskGuard, TaskSpawner, TokioTaskExecutor};
 use reth_transaction_pool::{noop::NoopTransactionPool, TransactionPool};
@@ -198,16 +198,14 @@ pub mod error;
 
 /// Eth utils
 pub mod eth;
-pub use eth::{
-    EthApiBuilderCtx, EthConfig, EthHandlers, FeeHistoryCacheBuilder, GasPriceOracleBuilder,
-};
+pub use eth::{EthApiBuilder, EthHandlers};
 
 // Rpc server metrics
 mod metrics;
 
 /// Convenience function for starting a server in one step.
 #[allow(clippy::too_many_arguments)]
-pub async fn launch<Provider, Pool, Network, Tasks, Events, EvmConfig, EthApi, EthApiB>(
+pub async fn launch<Provider, Pool, Network, Tasks, Events, EvmConfig, EthApi>(
     provider: Provider,
     pool: Pool,
     network: Network,
@@ -216,7 +214,7 @@ pub async fn launch<Provider, Pool, Network, Tasks, Events, EvmConfig, EthApi, E
     executor: Tasks,
     events: Events,
     evm_config: EvmConfig,
-    eth: EthApiB,
+    eth: Box<dyn EthApiBuilder<Provider, Pool, EvmConfig, Network, Tasks, Events, Output = EthApi>>,
 ) -> Result<RpcServerHandle, RpcError>
 where
     Provider: FullRpcProvider + AccountReader + ChangeSetReader,
@@ -225,8 +223,6 @@ where
     Tasks: TaskSpawner + Clone + 'static,
     Events: CanonStateSubscriptions + Clone + 'static,
     EvmConfig: ConfigureEvm,
-    EthApiB: FnOnce(&EthApiBuilderCtx<Provider, Pool, EvmConfig, Network, Tasks, Events>) -> EthApi
-        + 'static,
     EthApi: FullEthApiServer,
 {
     let module_config = module_config.into();
@@ -421,11 +417,13 @@ where
     /// also configures the auth (engine api) server, which exposes a subset of the `eth_`
     /// namespace.
     #[allow(clippy::type_complexity)]
-    pub fn build_with_auth_server<EngineApi, EngineT, EthApi, EthApiB>(
+    pub fn build_with_auth_server<EngineApi, EngineT, EthApi>(
         self,
         module_config: TransportRpcModuleConfig,
         engine: EngineApi,
-        eth: EthApiB,
+        eth: Box<
+            dyn EthApiBuilder<Provider, Pool, EvmConfig, Network, Tasks, Events, Output = EthApi>,
+        >,
     ) -> (
         TransportRpcModules,
         AuthRpcModule,
@@ -434,8 +432,6 @@ where
     where
         EngineT: EngineTypes + 'static,
         EngineApi: EngineApiServer<EngineT>,
-        EthApiB: FnOnce(&EthApiBuilderCtx<Provider, Pool, EvmConfig, Network, Tasks, Events>) -> EthApi
-            + 'static,
         EthApi: FullEthApiServer,
     {
         let Self { provider, pool, network, executor, events, evm_config } = self;
@@ -481,14 +477,15 @@ where
     ///     let eth_api = registry.eth_api();
     /// }
     /// ```
-    pub fn into_registry<EthApi, EthApiB>(
+    pub fn into_registry<EthApi>(
         self,
         config: RpcModuleConfig,
-        eth: EthApiB,
+        eth: Box<
+            dyn EthApiBuilder<Provider, Pool, EvmConfig, Network, Tasks, Events, Output = EthApi>,
+        >,
     ) -> RpcRegistryInner<Provider, Pool, Network, Tasks, Events, EthApi>
     where
-        EthApiB: FnOnce(&EthApiBuilderCtx<Provider, Pool, EvmConfig, Network, Tasks, Events>) -> EthApi
-            + 'static,
+        EthApi: 'static,
     {
         let Self { provider, pool, network, executor, events, evm_config } = self;
         RpcRegistryInner::new(provider, pool, network, executor, events, config, evm_config, eth)
@@ -496,14 +493,14 @@ where
 
     /// Configures all [`RpcModule`]s specific to the given [`TransportRpcModuleConfig`] which can
     /// be used to start the transport server(s).
-    pub fn build<EthApi, EthApiB>(
+    pub fn build<EthApi>(
         self,
         module_config: TransportRpcModuleConfig,
-        eth: EthApiB,
+        eth: Box<
+            dyn EthApiBuilder<Provider, Pool, EvmConfig, Network, Tasks, Events, Output = EthApi>,
+        >,
     ) -> TransportRpcModules<()>
     where
-        EthApiB: FnOnce(&EthApiBuilderCtx<Provider, Pool, EvmConfig, Network, Tasks, Events>) -> EthApi
-            + 'static,
         EthApi: FullEthApiServer,
     {
         let mut modules = TransportRpcModules::default();
@@ -631,13 +628,14 @@ impl<Provider, Pool, Network, Tasks, Events, EthApi>
 where
     Provider: StateProviderFactory + BlockReader + EvmEnvProvider + Clone + Unpin + 'static,
     Pool: Send + Sync + Clone + 'static,
-    Network: Clone,
-    Events: CanonStateSubscriptions + Clone,
+    Network: Clone + 'static,
+    Events: CanonStateSubscriptions + Clone + 'static,
     Tasks: TaskSpawner + Clone + 'static,
+    EthApi: 'static,
 {
     /// Creates a new, empty instance.
     #[allow(clippy::too_many_arguments)]
-    pub fn new<EvmConfig, EthApiB>(
+    pub fn new<EvmConfig>(
         provider: Provider,
         pool: Pool,
         network: Network,
@@ -645,12 +643,12 @@ where
         events: Events,
         config: RpcModuleConfig,
         evm_config: EvmConfig,
-        eth_api_builder: EthApiB,
+        eth_api_builder: Box<
+            dyn EthApiBuilder<Provider, Pool, EvmConfig, Network, Tasks, Events, Output = EthApi>,
+        >,
     ) -> Self
     where
         EvmConfig: ConfigureEvm,
-        EthApiB: FnOnce(&EthApiBuilderCtx<Provider, Pool, EvmConfig, Network, Tasks, Events>) -> EthApi
-            + 'static,
     {
         let blocking_pool_guard = BlockingTaskGuard::new(config.eth.max_tracing_requests);
