+++ crates/rpc/rpc-eth-types/src/builder/mod.rs
@@ -1,0 +1,4 @@
+//! `eth` namespace API builder types.
+
+pub mod config;
+pub mod ctx;
