+++ crates/rpc/rpc/Cargo.toml
@@ -33,6 +33,7 @@ reth-evm.workspace = true
 reth-rpc-eth-types.workspace = true
 reth-rpc-server-types.workspace = true
 reth-evm-optimism = { workspace = true, optional = true }
+reth-node-api.workspace = true
 
 # eth
 alloy-dyn-abi.workspace = true
