+++ crates/rpc/rpc/src/eth/filter.rs
@@ -16,8 +16,7 @@ use reth_primitives::{IntoRecoveredTransaction, TxHash};
 use reth_provider::{BlockIdReader, BlockReader, EvmEnvProvider, ProviderError};
 use reth_rpc_eth_api::EthFilterApiServer;
 use reth_rpc_eth_types::{
-    logs_utils::{self, append_matching_block_logs},
-    EthApiError, EthFilterError, EthStateCache, EthSubscriptionIdProvider,
+    logs_utils::{self, append_matching_block_logs}, EthApiError, EthFilterConfig, EthFilterError, EthStateCache, EthSubscriptionIdProvider
 };
 use reth_rpc_server_types::ToRpcResult;
 use reth_rpc_types::{
@@ -512,56 +511,6 @@ where
         }
 
         Ok(all_logs)
-    }
-}
-
-/// Config for the filter
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct EthFilterConfig {
-    /// Maximum number of blocks that a filter can scan for logs.
-    ///
-    /// If `None` then no limit is enforced.
-    pub max_blocks_per_filter: Option<u64>,
-    /// Maximum number of logs that can be returned in a single response in `eth_getLogs` calls.
-    ///
-    /// If `None` then no limit is enforced.
-    pub max_logs_per_response: Option<usize>,
-    /// How long a filter remains valid after the last poll.
-    ///
-    /// A filter is considered stale if it has not been polled for longer than this duration and
-    /// will be removed.
-    pub stale_filter_ttl: Duration,
-}
-
-impl EthFilterConfig {
-    /// Sets the maximum number of blocks that a filter can scan for logs.
-    pub const fn max_blocks_per_filter(mut self, num: u64) -> Self {
-        self.max_blocks_per_filter = Some(num);
-        self
-    }
-
-    /// Sets the maximum number of logs that can be returned in a single response in `eth_getLogs`
-    /// calls.
-    pub const fn max_logs_per_response(mut self, num: usize) -> Self {
-        self.max_logs_per_response = Some(num);
-        self
-    }
-
-    /// Sets how long a filter remains valid after the last poll before it will be removed.
-    pub const fn stale_filter_ttl(mut self, duration: Duration) -> Self {
-        self.stale_filter_ttl = duration;
-        self
-    }
-}
-
-impl Default for EthFilterConfig {
-    fn default() -> Self {
-        Self {
-            max_blocks_per_filter: None,
-            max_logs_per_response: None,
-            // 5min
-            stale_filter_ttl: Duration::from_secs(5 * 60),
-        }
     }
 }
