+++ crates/rpc/rpc/src/eth/mod.rs
@@ -9,7 +9,7 @@ pub mod pubsub;
 /// Implementation of `eth` namespace API.
 pub use bundle::EthBundle;
 pub use core::EthApi;
-pub use filter::{EthFilter, EthFilterConfig};
+pub use filter::EthFilter;
 pub use pubsub::EthPubSub;
 
 pub use helpers::signer::DevSigner;
